requirements.txt
Django==4.1.2
environs==9.5.0
dj-database-url==1.0.0
psycopg2-binary==2.9.4
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
loguru==0.6.0
django-simple-history==3.1.1
django-debug-toolbar
channels==3.0.5
daphne==3.0.2
channels-redis==4.0.0
django-eventstream==4.4.1
celery==5.2.7
django-celery-beat==2.4.0
django_cors_headers==3.13.0
Pillow==9.3.0
httpx==0.23.0
django-constance==2.9.1
